reorgonize utility functions
